Add layer to retrieve book by Isbn
This adds a layer on top of the lookup modules which allows to fake out
the interaction with the external API in the way J.Valim mentioned
recently. It's an interesting approach. I'm also still missing the test
for the controller as I'm not sure how to write controller tests yet.